PHP code audit for the com.aptana.editor.php plugin(first phase)

+++ plugins/com.aptana.editor.php/src/com/aptana/editor/php/internal/parser/PHTMLTokens.java
@@ -41,7 +41,7 @@ public class PHTMLTokens
 		{
 			return NAME_UNKNOWN;
 		}
-		return NAMES[token];
+		return NAMES[(int)token];
 	}
 
 	public static short getToken(String tokenName)
